app_v3.py
import gradio as gr
import torch
from torchvision import models
import torchvision.transforms as T
from PIL import Image, ImageFilter, ImageOps, ImageEnhance
import numpy as np
from pytorch_grad_cam import GradCAM
from pytorch_grad_cam.utils.image import show_cam_on_image
import requests
import traceback

print(">>> 正在初始化模型 (ResNet101)...")
seg_model = models.segmentation.deeplabv3_resnet101(weights='DEFAULT').eval()
cls_model = models.resnet101(weights='DEFAULT').eval()

try:
    resp = requests.get("https://raw.githubusercontent.com/pytorch/hub/master/imagenet_classes.txt")
    imagenet_labels = resp.text.split("\n")
except:
    imagenet_labels = ["Label Error"] * 1000

COCO_CLASSES = [
    '__background__', 'aeroplane', 'bicycle', 'bird', 'boat', 'bottle', 'bus',
    'car', 'cat', 'chair', 'cow', 'diningtable', 'dog', 'horse', 'motorbike',
    'person', 'pottedplant', 'sheep', 'sofa', 'train', 'tvmonitor'
]

def preprocess_cls(img):
    t = T.Compose([
        T.Resize((224, 224)),
        T.ToTensor(),
        T.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
    ])
    return t(img).unsqueeze(0)

def preprocess_seg(img):
    t = T.Compose([
        T.ToTensor(),
        T.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
    ])
    return t(img).unsqueeze(0)

def predict_top_k_readable(model, img_pil, k=3):
    tensor = preprocess_cls(img_pil)
    with torch.no_grad():
        out = model(tensor)
    prob = torch.nn.functional.softmax(out[0], dim=0)
    topk_prob, topk_id = torch.topk(prob, k)
    
    best_label = imagenet_labels[topk_id[0].item()]
    
    html_content = ""
    total_conf = 0
    for i in range(k):
        label = imagenet_labels[topk_id[i].item()]
        score = topk_prob[i].item()
        total_conf += score
        
        bar_length = int(score * 100)
        color = "#4ade80" if i == 0 else "#94a3b8"
        
        html_content += f"""
        <div style="margin-bottom: 8px;">
            <div style="display: flex; justify-content: space-between; margin-bottom: 2px;">
                <span style="font-weight: 500;">{label.title()}</span>
                <span>{score:.1%}</span>
            </div>
            <div style="background-color: #e2e8f0; border-radius: 9999px; height: 8px; width: 100%;">
                <div style="background-color: {color}; height: 8px; border-radius: 9999px; width: {bar_length}%;"></div>
            </div>
        </div>
        """
        
    return best_label, html_content, total_conf

def generate_cam_resized(model, img_pil, target_size):
    target_layers = [model.layer4[-1]]
    cam = GradCAM(model=model, target_layers=target_layers)
    input_tensor = preprocess_cls(img_pil)
    
    grayscale_cam = cam(input_tensor=input_tensor, targets=None)[0, :]
    
    img_small = img_pil.resize((224, 224))
    rgb_img = np.float32(img_small) / 255
    visualization = show_cam_on_image(rgb_img, grayscale_cam, use_rgb=True)
    
    vis_pil = Image.fromarray(visualization).resize(target_size, resample=Image.BICUBIC)
    
    return vis_pil

def fix_holes_ultimate(img, mask_pil, padding=10):
    mask_fat = mask_pil.filter(ImageFilter.MaxFilter(21)) 
    
    mask_soft = mask_fat.filter(ImageFilter.GaussianBlur(radius=5))
    
    patch_layer = img.filter(ImageFilter.GaussianBlur(radius=40))
    
    bg_white = Image.new('RGB', img.size, (255, 255, 255))
    filled_hole_img = Image.composite(patch_layer, bg_white, mask_soft)
    final_composite = Image.composite(img, filled_hole_img, mask_pil)

    mask_arr = np.array(mask_fat) 
    rows = np.any(mask_arr, axis=1)
    cols = np.any(mask_arr, axis=0)
    if not np.any(rows) or not np.any(cols): return final_composite
    
    ymin, ymax = np.where(rows)[0][[0, -1]]
    xmin, xmax = np.where(cols)[0][[0, -1]]
    
    w, h = img.size
    xmin = max(0, xmin - padding)
    xmax = min(w, xmax + padding)
    ymin = max(0, ymin - padding)
    ymax = min(h, ymax + padding)
    
    cropped = final_composite.crop((xmin, ymin, xmax, ymax))
    
    w_crop, h_crop = cropped.size
    max_side = max(w_crop, h_crop)
    square_img = Image.new('RGB', (max_side, max_side), (255, 255, 255))
    offset_x = (max_side - w_crop) // 2
    offset_y = (max_side - h_crop) // 2
    square_img.paste(cropped, (offset_x, offset_y))
    
    enhancer = ImageEnhance.Contrast(square_img)
    square_img = enhancer.enhance(1.15)
    
    return square_img

def inference_pipeline(original_image, target_class):
    if original_image is None: return [None]*4 + ["请先上传图片"]
    
    try:
        original_image = original_image.convert("RGB")
        orig_size = original_image.size
        
        orig_label, orig_html, _ = predict_top_k_readable(cls_model, original_image, k=3)
        heatmap_orig = generate_cam_resized(cls_model, original_image, orig_size)
        
        input_tensor = preprocess_seg(original_image)
        with torch.no_grad(): output = seg_model(input_tensor)['out'][0]
        predictions = output.argmax(0)
        
        try: idx = COCO_CLASSES.index(target_class)
        except: return [None]*4 + [f"类别不支持"]
            
        mask_np = (predictions == idx).cpu().numpy().astype(np.uint8)
        mask_pil = Image.fromarray(mask_np * 255)
        if mask_pil.size != original_image.size:
            mask_pil = mask_pil.resize(original_image.size, resample=Image.NEAREST)
        
        if np.array(mask_pil).sum() == 0:
            return [None]*4 + [f"⚠️ 在图中未检测到 {target_class}。<br>建议：1. 检查目标是否在 20 类名单中；2. 换一张主体更清晰的图片。"]

        img_final = fix_holes_ultimate(original_image, mask_pil, padding=20)
        
        focus_label, focus_html, total_conf = predict_top_k_readable(cls_model, img_final, k=3)
        heatmap_final = generate_cam_resized(cls_model, img_final, img_final.size)
        
        report_html = f"""
        <div style="padding: 20px; background-color: #f8fafc; border-radius: 10px; border: 1px solid #e2e8f0;">
            <h3 style="margin-top: 0; color: #1e293b;">🕵️‍♂️ 视觉注意力迁移报告</h3>
            
            <div style="display: flex; gap: 20px; flex-wrap: wrap;">
                
                <div style="flex: 1; min-width: 250px; background: white; padding: 15px; border-radius: 8px; box-shadow: 0 1px 3px rgba(0,0,0,0.1);">
                    <h4 style="margin: 0 0 10px 0; color: #64748b;">🚫 原始环境 (干扰严重)</h4>
                    <div style="font-size: 14px; margin-bottom: 5px;">AI 判定为: <b>{orig_label}</b></div>
                    {orig_html}
                </div>

                <div style="flex: 1; min-width: 250px; background: white; padding: 15px; border-radius: 8px; box-shadow: 0 4px 6px rgba(0,0,0,0.1); border-left: 5px solid #4ade80;">
                    <h4 style="margin: 0 0 10px 0; color: #166534;">✅ 聚焦优化后 (纠正成功)</h4>
                    <div style="font-size: 14px; margin-bottom: 5px;">AI 判定为: <b>{focus_label}</b></div>
                    {focus_html}
                    <div style="margin-top: 10px; padding-top: 10px; border-top: 1px dashed #cbd5e1; font-size: 13px; color: #475569;">
                        🔥 <b>聚合置信度: {total_conf:.1%}</b><br>
                        (对该目标的确认程度)
                    </div>
                </div>
                
            </div>
            
            <p style="margin-top: 15px; font-size: 13px; color: #64748b; line-height: 1.5;">
                💡 <b>技术解析:</b> 系统检测到原始环境中存在背景干扰（如共现物体），导致模型注意力分散。通过 <b>强力补洞算法</b> 与 <b>白底证件照模式</b>，我们成功修复了目标掩码的缺失，强制模型将注意力 100% 集中在目标主体上。
            </p>
        </div>
        """
        
        return mask_pil, heatmap_orig, img_final, heatmap_final, report_html

    except Exception as e:
        error_msg = traceback.format_exc()
        print("❌ 发生错误:", error_msg)
        return [None]*4 + [f"<div style='color: red;'><h3>🚫 程序运行出错</h3><p>错误信息: {e}</p><p>请检查后台终端查看详细报错。</p></div>"]

with gr.Blocks(title="AI Vision Pro V9", theme=gr.themes.Soft()) as demo:
    gr.Markdown("# 👁️ 深度学习视觉注意力引导系统 (V9.3)")
    
    with gr.Row():
        with gr.Column(scale=1):
            img_input = gr.Image(type="pil", label="📸 上传图片")
            cls_selector = gr.Dropdown(
                choices=[
                    "cat", "dog", "person", "horse", "bird", "sheep", "cow", 
                    "aeroplane", "bicycle", "boat", "bus", "car", "motorbike", "train", 
                    "bottle", "chair", "diningtable", "pottedplant", "sofa", "tvmonitor"
                ], 
                value="cat", 
                label="🎯 目标对象"
            )
            btn_run = gr.Button("✨ 启动对比分析", variant="primary")
    
    gr.Markdown("---")
    
    with gr.Row():
        with gr.Column():
            gr.Markdown("### 1️⃣ 原始状态")
            out_cam_orig = gr.Image(label="原始热力图 ", type="pil")
        with gr.Column():
            gr.Markdown("### 2️⃣ 中间过程")
            out_mask = gr.Image(label="语义分割掩码", type="pil")
        with gr.Column():
            gr.Markdown("### 3️⃣ 最终状态")
            out_focus = gr.Image(label="修复后证件照 ", type="pil")
        with gr.Column():
            gr.Markdown("### 4️⃣ 最终验证")
            out_cam_final = gr.Image(label="最终热力图 ", type="pil")

    out_report = gr.HTML(label="分析报告")

    btn_run.click(fn=inference_pipeline, inputs=[img_input, cls_selector], 
                  outputs=[out_mask, out_cam_orig, out_focus, out_cam_final, out_report])

if __name__ == "__main__":
    print(">>> 启动成功！请在浏览器打开链接")
    demo.launch(share=False)

